requirements.txt
pymongo==4.1.1
redis==3.5.3
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
PyJWT==2.1.0
requests==2.26.0
user-agents

